+++ internal/cluster/pipeline/service.go
@@ -8,32 +8,32 @@ import (
 	"github.com/myntra/pipeline"
 )
 
-// Deployment will implement step interface for Deployments
-type Deployment struct {
+// Service will implement step interface for Services
+type Service struct {
 	pipeline.StepContext
 	client *discovery.Client
 	broker broker.Handler
 }
 
-// NewDeployment - constructor
-func NewDeployment(client *discovery.Client, broker broker.Handler) *Deployment {
-	return &Deployment{
+// NewService - constructor
+func NewService(client *discovery.Client, broker broker.Handler) *Service {
+	return &Service{
 		client: client,
 		broker: broker,
 	}
 }
 
 // Exec - step interface
-func (d *Deployment) Exec(request *pipeline.Request) *pipeline.Result {
+func (d *Service) Exec(request *pipeline.Request) *pipeline.Result {
 	// it will contain a pipeline to run
-	log.Println("Deployment Discovery Started")
+	log.Println("Service Discovery Started")
 
 	// get all namespaces
 	namespaces := NamespaceName
 
 	for _, namespace := range namespaces {
-		// get Deployments
-		deployments, err := d.client.ListDeployments(namespace)
+		// get Services
+		services, err := d.client.ListServices(namespace)
 		if err != nil {
 			return &pipeline.Result{
 				Error: err,
@@ -41,16 +41,15 @@ func (d *Deployment) Exec(request *pipeline.Request) *pipeline.Result {
 		}
 
 		// processing
-		for _, deployment := range deployments {
-			// publishing discovered deployment
+		for _, service := range services {
+			// publishing discovered Service
 			err := d.broker.Publish(Subject, broker.Message{
-				Type:   "Deployment",
-				Object: deployment,
+				Object: service,
 			})
 			if err != nil {
-				log.Printf("Error publishing deployment named %s", deployment.Name)
+				log.Printf("Error publishing service named %s", service.Name)
 			} else {
-				log.Printf("Published deployment named %s", deployment.Name)
+				log.Printf("Published service named %s", service.Name)
 			}
 		}
 	}
@@ -62,7 +61,7 @@ func (d *Deployment) Exec(request *pipeline.Request) *pipeline.Result {
 }
 
 // Cancel - step interface
-func (d *Deployment) Cancel() error {
+func (d *Service) Cancel() error {
 	d.Status("cancel step")
 	return nil
 }
