+++ internal/meshes/istio/informers/service_entry.go
