+++ pkg/broker/broker.go
@@ -1,17 +1,20 @@
 package broker
 
-import (
-	"github.com/layer5io/meshsync/pkg/broker/nats"
+var (
+	List   ObjectType = "list"
+	Single ObjectType = "single"
 )
 
+type ObjectType string
+
 type Message struct {
-	Type   string
+	Type   ObjectType
 	Object interface{}
 }
 
 type PublishInterface interface {
-	Publish(string, interface{}) error
-	PublishWithCallback(string, string, interface{}) error
+	Publish(string, Message) error
+	PublishWithCallback(string, string, Message) error
 }
 
 type SubscribeInterface interface {
@@ -22,20 +25,4 @@ type SubscribeInterface interface {
 type Handler interface {
 	PublishInterface
 	SubscribeInterface
-}
-
-const (
-	NATSKey = "nats"
-)
-
-func New(kind, url string) (Handler, error) {
-	var broker Handler
-	// switch kind {
-	// case NATSKey:
-	// 	return nats.New(url)
-	// }
-	if kind == NATSKey {
-		return nats.New(url)
-	}
-	return broker, nil
 }
